Content Update
I added some more stuff to the front page of the website, with 4 photos and one video. Made it look better with borders and removed some bad comments and other stuff.

The photos are not mine, but the users are credited in both the text and the file names. I'll take them down if they wish so.