Text Updates, and Small Changes
Removed the McDonalds bit, since I got it wrong. Technically, Boid Gaming is in the Amino Hub, not the Galactic Hub, something I somehow managed to miss.

I added a small list of farms, and changed the CSS a bit. This website looks a little funky on large monitors, so I'm working on a solution to that. The website looks fine on phones and smaller monitors.