CSS Changes, New Page, New Links
Added a couple of small changes tonight, from this mornings revision.

I added an ID to each <main> of the index.html page, and linked to it near the top of the page, so you could jump to each section. It simply uses <href = "#id"> and it jumps to that section.

I added some small padding to article text and article list classes so that they wouldn't touch the edges of the pages. This was a QoL suggestion from a user on the discord server.

I added a small framework for a new page, but it has zero content, so it serves no purpose right now.